Split stage artefact tests by concern
Move the monolithic `test_stage.py` coverage into focused modules for
environment, config, resolution, output, and pipeline behaviour.

Extract shared staging config and PowerShell workspace helpers so the
pipeline tests keep the same assertions with less duplicated setup.
